Move ImageBlock to components/ui
