--- airbyte-webapp/src/components/ImageBlock/index.tsx
@@ -1,5 +1,0 @@
-import ImageBlock from "./ImageBlock";
-
-export * from "./ImageBlock";
-
-export default ImageBlock;
