+++ airbyte-webapp/src/components/ui/ImageBlock/ImageBlock.module.scss
@@ -1,5 +1,5 @@
-@use "../../scss/colors";
-@use "../../scss/fonts";
+@use "../../../scss/colors";
+@use "../../../scss/fonts";
 
 .iconContainer {
   height: 25px;
