+++ airbyte-webapp/src/components/ui/ImageBlock/ImageBlock.stories.tsx
@@ -3,15 +3,15 @@ import { ComponentStory, ComponentMeta } from "@storybook/react";
 import { ImageBlock } from "./ImageBlock";
 
 export default {
-  title: "Ui/ImageBlock",
+  title: "UI/ImageBlock",
   component: ImageBlock,
   argTypes: {},
 } as ComponentMeta<typeof ImageBlock>;
 
 const Template: ComponentStory<typeof ImageBlock> = (args) => <ImageBlock {...args} />;
 
-export const ImageBlockControl = Template.bind({});
-ImageBlockControl.args = {
+export const Primary = Template.bind({});
+Primary.args = {
   img: undefined,
   num: undefined,
   small: undefined,
