+++ airbyte-webapp/src/components/ui/ImageBlock/ImageBlock.tsx
