+++ airbyte-webapp/src/components/ui/ImageBlock/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ImageBlock";
